api_requirements.txt
fastapi
uvicorn
python-telegram-bot
requests
python-dotenv
pydantic
